+++ src/creation/index.ts
@@ -4,12 +4,14 @@ import * as vscode from "vscode";
 import * as config from "../config";
 import { EXTENSION_NAME, SWING_FILE } from "../constants";
 import { DEFAULT_MANIFEST, openSwing } from "../preview";
+import { store } from "../store";
 import { stringToByteArray, withProgress } from "../utils";
 import {
   enableGalleries,
   loadGalleries,
   registerTemplateProvider
 } from "./galleryProvider";
+import { initializeStorage, storage } from "./storage";
 
 export interface SwingFile {
   filename: string;
@@ -20,15 +22,21 @@ interface CodeSwingTemplateItem extends vscode.QuickPickItem {
   files?: SwingFile[];
 }
 
-export async function newSwing(
-  uri: vscode.Uri | ((files: SwingFile[]) => Promise<vscode.Uri>),
-  title: string = "Create new swing"
-) {
-  const quickPick = vscode.window.createQuickPick();
-  quickPick.title = title;
-  quickPick.placeholder = "Select the swing template to use";
-  quickPick.matchOnDescription = true;
+async function createSwingDirectory() {
+  const scratchDirectory =
+          config.get("scratchDirectory") ||
+          path.join(os.tmpdir(), EXTENSION_NAME);
+  const dayjs = require("dayjs");
+  const timestamp = dayjs().format("YYYY-MM-DD (hh-mm-ss A)");
+  const swingDirectory = path.join(scratchDirectory, timestamp);
 
+  const uri = vscode.Uri.file(swingDirectory);
+  await vscode.workspace.fs.createDirectory(uri);
+
+  return uri;
+}
+
+async function getTemplates(): Promise<CodeSwingTemplateItem[]> {
   const galleries = await loadGalleries();
 
   const templates: CodeSwingTemplateItem[] = galleries
@@ -37,13 +45,40 @@ export async function newSwing(
     .sort((a, b) => a.title.localeCompare(b.title))
     .map((t) => ({ ...t, label: t.title }));
 
+  return templates;
+}
+
+export async function newSwing(
+  uri: vscode.Uri | ((files: SwingFile[]) => Promise<vscode.Uri>),
+  title: string = "Create new swing"
+) {
+  const quickPick = vscode.window.createQuickPick();
+  quickPick.title = title;
+  quickPick.placeholder = "Select the template to use";
+  quickPick.matchOnDescription = true;
+  quickPick.ignoreFocusOut = true;
+  
+  const templates = await getTemplates();
   if (templates.length === 0) {
     templates.push({
       label:
         "No templates available. Configure your template galleries and try again.",
     });
   }
-
+  
+  const mru = storage.getTemplateMRU();
+  if (mru && mru.length > 0) {
+    for (let i = mru.length - 1; i >= 0; i--) {
+      const itemIndex = templates.findIndex(gallery => gallery.label === mru[i]);
+      if (itemIndex !== -1) {
+        const [item] = templates.splice(itemIndex, 1);
+        item.alwaysShow = true;
+        item.description = "Recently used";
+        templates.unshift(item);
+      } 
+    }
+  }
+  
   quickPick.items = templates;
   quickPick.buttons = [
     {
@@ -61,11 +96,11 @@ export async function newSwing(
 
     const template = quickPick.selectedItems[0] as CodeSwingTemplateItem;
     if (template.files) {
-      const swingUri = await withProgress("Creating swing...", async () =>
+      await withProgress("Creating swing...", async () =>
         newSwingFromTemplate(template.files!, uri)
       );
 
-      openSwing(swingUri);
+      await storage.addTemplateToMRU(template.label);
     }
   });
 
@@ -75,7 +110,7 @@ export async function newSwing(
 async function newSwingFromTemplate(
   files: SwingFile[],
   uri: vscode.Uri | ((files: SwingFile[]) => Promise<vscode.Uri>)
-): Promise<vscode.Uri> {
+) {
   const manifest = files.find((file) => file.filename === SWING_FILE)
   if (!manifest) {
     const content = JSON.stringify(DEFAULT_MANIFEST, null, 2);
@@ -92,8 +127,9 @@ async function newSwingFromTemplate(
     }
   }
 
+  let swingUri: vscode.Uri;
   if (uri instanceof Function) {
-    return uri(
+    swingUri = await uri(
       files.map(({ filename, content }) => ({
         filename,
         content: content ? content : "",
@@ -109,8 +145,10 @@ async function newSwingFromTemplate(
         );
       })
     );
-    return uri;
+    swingUri = uri;
   }
+
+  openSwing(swingUri);
 }
 
 async function promptForGalleryConfiguration(
@@ -126,7 +164,7 @@ async function promptForGalleryConfiguration(
   const galleries = (await loadGalleries())
     .sort((a, b) => a.title.localeCompare(b.title))
     .map((gallery) => ({ ...gallery, label: gallery.title }));
-
+  
   quickPick.items = galleries;
   quickPick.selectedItems = galleries.filter((gallery) => gallery.enabled);
 
@@ -152,32 +190,35 @@ async function promptForGalleryConfiguration(
   quickPick.show();
 }
 
-export async function registerCreationModule(
+export function registerCreationModule(
   context: vscode.ExtensionContext,
-  api: any
+  api: any,
+  syncKeys: string[]
 ) {
   context.subscriptions.push(
     vscode.commands.registerCommand(
-      `${EXTENSION_NAME}.newScratchSwing`,
+      `${EXTENSION_NAME}.newSwing`,
       async () => {
-        const scratchDirectory =
-          config.get("scratchDirectory") ||
-          path.join(os.tmpdir(), EXTENSION_NAME);
-
-        const dayjs = require("dayjs");
-        const timestamp = dayjs().format("YYYY-MM-DD (hh-mm-ss A)");
-        const swingDirectory = path.join(scratchDirectory, timestamp);
-
-        const uri = vscode.Uri.file(swingDirectory);
-        await vscode.workspace.fs.createDirectory(uri);
-
+        const uri = await createSwingDirectory();
         newSwing(uri);
       }
     )
   );
 
   context.subscriptions.push(
-    vscode.commands.registerCommand(`${EXTENSION_NAME}.newSwing`, async () => {
+    vscode.commands.registerCommand(`${EXTENSION_NAME}.newSwingFromLastTemplate`, async () => {
+      const [latestTemplate] = storage.getTemplateMRU();
+      const templates = await getTemplates();
+      const template = templates.find(template => template.label === latestTemplate);
+      if (template) {
+        const uri = await createSwingDirectory();
+        newSwingFromTemplate(template.files!, uri);
+      }
+    })
+  );
+
+  context.subscriptions.push(
+    vscode.commands.registerCommand(`${EXTENSION_NAME}.newSwingDirectory`, async () => {
       const folder = await vscode.window.showOpenDialog({
         canSelectFolders: true,
         canSelectFiles: false,
@@ -191,6 +232,34 @@ export async function registerCreationModule(
     })
   );
 
+  context.subscriptions.push(
+    vscode.commands.registerCommand(`${EXTENSION_NAME}.saveCurrentSwing`, async () => {
+      const folder = await vscode.window.showOpenDialog({
+        canSelectFolders: true,
+        canSelectFiles: false,
+        canSelectMany: false,
+        defaultUri: vscode.workspace.workspaceFolders?.[0].uri,
+      });
+
+      if (!folder) {
+        return;
+      }
+
+      await withProgress("Saving swing...", async () => {
+        const files = await vscode.workspace.fs.readDirectory(store.activeSwing!.rootUri);
+        return Promise.all(files.map(async ([file]) => {
+          const sourceUri = vscode.Uri.joinPath(store.activeSwing!.rootUri, file);
+          const contents = await vscode.workspace.fs.readFile(sourceUri);
+
+          const uri = vscode.Uri.joinPath(folder[0], file);
+          await vscode.workspace.fs.writeFile(uri, contents);
+        }));
+      });
+    })
+  );
+
+  initializeStorage(context, syncKeys);
+  
   api.newSwing = newSwing;
   api.registerTemplateProvider = registerTemplateProvider;
 }