+++ src/preview/tutorials/index.ts
@@ -2,7 +2,7 @@ import { ExtensionContext } from "vscode";
 import { registerInputFileSystemProvider } from "./inputFileSystem";
 import { initializeStorage } from "./storage";
 
-export async function registerTutorialModule(context: ExtensionContext) {
+export async function registerTutorialModule(context: ExtensionContext, syncKeys: string[]) {
   registerInputFileSystemProvider();
-  initializeStorage(context);
+  initializeStorage(context, syncKeys);
 }